+++ src/main/java/org/jenkinsci/plugins/artifactdeployer/ArtifactDeployerPublisher.java
@@ -178,21 +178,23 @@ public class ArtifactDeployerPublisher extends Recorder implements Serializable 
                             //Delete output
                             if (entry.isDeleteRemoteArtifacts()) {
                                 List<ArtifactDeployerVO> listArtifacts = info.get(entry.getUniqueId());
-                                for (ArtifactDeployerVO vo : listArtifacts) {
-                                    FilePath remoteArtifactPath = new FilePath(build.getWorkspace().getChannel(), vo.getRemotePath());
-                                    try {
-                                        if (remoteArtifactPath.exists()) {
-                                            remoteArtifactPath.deleteRecursive();
-                                        }
+                                if (listArtifacts != null) {
+                                    for (ArtifactDeployerVO vo : listArtifacts) {
+                                        FilePath remoteArtifactPath = new FilePath(build.getWorkspace().getChannel(), vo.getRemotePath());
+                                        try {
+                                            if (remoteArtifactPath.exists()) {
+                                                remoteArtifactPath.deleteRecursive();
+                                            }
 
-                                        if (remoteArtifactPath.getParent().exists() && remoteArtifactPath.getParent().list().size() == 0) {
-                                            remoteArtifactPath.getParent().delete();
-                                        }
+                                            if (remoteArtifactPath.getParent().exists() && remoteArtifactPath.getParent().list().size() == 0) {
+                                                remoteArtifactPath.getParent().delete();
+                                            }
 
-                                    } catch (IOException ioe) {
-                                        logger.log(Level.SEVERE, "Error when deleting artifacts.", ioe);
-                                    } catch (InterruptedException ie) {
-                                        logger.log(Level.SEVERE, "Error when deleting artifacts.", ie);
+                                        } catch (IOException ioe) {
+                                            logger.log(Level.SEVERE, "Error when deleting artifacts.", ioe);
+                                        } catch (InterruptedException ie) {
+                                            logger.log(Level.SEVERE, "Error when deleting artifacts.", ie);
+                                        }
                                     }
                                 }
